requirements.txt
# 모듈명, 버전 설치시 꼭 추가! 은지님 화이팅
# 예시) tensorflow==2.7.0
apache-airflow==2.5.1
pendulum==2.0.5
pymysql==1.1.0
prettytable==3.7.0
selenium==4.11.2
konlpy==0.6.0
boto3==1.26.51
beautifulsoup4==4.11.1
pandas==1.3.5
requests==2.28.2
webdriver-manager==4.0.1
apache-airflow-providers-mysql==4.0.0
apache-airflow-providers-slack==7.2.0
apache-airflow-providers-amazon==7.1.0
